+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 23. Add some flexibility to `CommonDropdownMenu`, namely the ability to define anchor and transform origins and styling
     parameters.
 24. Add `isLoading` to `CommonSelect` and `CommonMultiSelect`, which adds an indefinite loading bar beneath the select when true.
+25. Remove default sidebar content and external domains. Should specify explicit content, such as from `@grail/common-private`.
 
 ## 1.0.0 - 2019-06-17